condVarRotated {lme4breeding}	R Documentation

Get the rotated conditional variance matrix

Description

Retrieve the conditional variance matrix from the random effects rotated by the Cholesky factors and Eigen factors if applicable.

Usage

condVarRotated(object)

Arguments

object
model object of class lmeb

Value

Matrix

Author(s)

Giovanny Covarrubias

See Also

image, lmeb

Examples

data(DT_example, package="enhancer")
DT <- DT_example
A <- A_example
head(DT)

## Compound simmetry (CS) model
ans1 <- lmeb(Yield~Env + (1|Name) + (1|Env:Name),
                 data=DT)
vc <- VarCorr(ans1); print(vc,comp=c("Variance"))

condVarMat <- condVarRotated(ans1)
# image(condVarMat)

[Package lme4breeding version 1.1.1 Index]