lmeb-class {lme4breeding}	R Documentation

Relationship-based mixed-effects model fits

Description

A mixed-effects model fit by lmeb. This class extends class "merMod" class and includes one additional slot, relfac, which is a list of (left) Cholesky factors of the relationship matrices derived from "lmeb" objects.

Objects from the Class

Objects are created by calls to the lmeb function.

Slots

relfac:

A list of relationship matrix factors. All other slots are inherited from class "merMod".

udu:

A list of eigen decomposition elements. All other slots are inherited from class "merMod".

Extends

Class "merMod", directly.

Methods

fitted

signature(object = "lmeb"): actually a non-method in that fitted doesn't apply to such objects because of the pre-whitening.

ranef

signature(object = "lmeb"): back-transforms BLUPs and their conditional variances when models include the relationship between levels of random effects as returned for the object viewed as a "merMod)" object.

residuals

signature(object = "lmeb"): also a non-method for the same reason as fitted

See Also

lmeb

Examples

showClass("lmeb")

data(DT_example, package="enhancer")
DT <- DT_example
A <- A_example

## Compound simmetry (CS) model
ans1 <- lmeb(Yield~Env + (1|Name) + (1|Env:Name),
                 data=DT)
                 
fitted(ans1)
residuals(ans1)
rr <- ranef(ans1)

[Package lme4breeding version 1.1.1 Index]