smm {lme4breeding}	R Documentation

sparse model matrix

Description

smm creates a sparse model matrix for the levels of the random effect to be used with the lmeb solver.

Usage

  smm(x)

Arguments

x
vector of observations for the random effect.

Value

$res

a model matrix for a given factor.

Author(s)

Giovanny Covarrubias-Pazaran

References

Giovanny Covarrubias-Pazaran (2024). lme4breeding: enabling genetic evaluation in the age of genomic data. To be submitted to Bioinformatics.

Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48.

See Also

The lmeb solver.

Examples

x <- as.factor(c(1:5,1:5,1:5));x
smm(x)

[Package lme4breeding version 1.1.1 Index]